Migrate a bunch of files/folders for prefab registry
