+++ src/ValheimRAFT.Unity/Assets/ValheimVehicles/SharedScripts/SharedScripts.Validation/SkipValidationAttribute.cs
@@ -1,0 +1,12 @@
+// ReSharper disable ArrangeNamespaceBody
+// ReSharper disable NamespaceStyle
+
+using System;
+
+namespace ValheimVehicles.SharedScripts.Validation
+{
+  [AttributeUsage(AttributeTargets.Field)]
+  public sealed class SkipValidationAttribute : Attribute
+  {
+  }
+}
