+++ src/ValheimRAFT/AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 1957239098
+CRC: 2585269695
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
