+++ src/ValheimVehicles/ValheimVehicles.Prefabs/ValheimVehicles.Prefabs.Registry/VehicleHammerRegistry.cs
@@ -1,0 +1,6 @@
+namespace Registry;
+
+public class VehicleHammerRegistry
+{
+  
+}
